requirements.txt
asgiref==3.7.2
black==23.3.0
boto3==1.33.3
botocore==1.33.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
cryptography==41.0.4
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.1.0
Django==4.2.3
django-cors-headers==4.3.0
django-dotenv==1.4.2
django-extensions==3.2.3
django-filter==23.3
django-import-export==3.3.1
django-unfold==0.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-nested==1.2.24
drf-nested-routers==0.93.4
drf-writable-nested==0.7.0
et-xmlfile==1.1.0
gunicorn==21.2.0
h11==0.14.0
idna==3.4
importlib-metadata==6.7.0
jmespath==1.0.1
MarkupPy==1.14
mypy-extensions==1.0.0
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pathspec==0.11.2
Pillow==10.1.0
platformdirs==3.11.0
psycopg==3.1.12
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.8.2
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.8.0
ua-parser==0.18.0
urllib3==2.0.7
user-agents==2.2.0
uvicorn==0.23.2
whitenoise==6.5.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.17.0
